+++ R/process.R
@@ -229,6 +229,10 @@ process <- R6::R6Class(
     #'   object is garbage collected.
     #' @param cleanup_tree Whether to kill the process and its child
     #'   process tree when the `process` object is garbage collected.
+    #' @param cleanup_grace Grace period between `SIGTERM` and `SIGKILL`.
+    #'   Only has an effect on Unix platforms. Set to 0 to terminate abruptly
+    #'   with `SIGKILL` only. Currently defaults to 0 until we implement
+    #'   a better approach on session quit.
     #' @param wd Working directory of the process. It must exist.
     #'   If `NULL`, then the current working directory is used.
     #' @param echo_cmd Whether to print the command to the screen before
@@ -275,6 +279,7 @@ process <- R6::R6Class(
       env = NULL,
       cleanup = TRUE,
       cleanup_tree = FALSE,
+      cleanup_grace = 0.0,
       wd = NULL,
       echo_cmd = FALSE,
       supervise = FALSE,
@@ -300,6 +305,7 @@ process <- R6::R6Class(
         env,
         cleanup,
         cleanup_tree,
+        cleanup_grace,
         wd,
         echo_cmd,
         supervise,
@@ -758,6 +764,7 @@ process <- R6::R6Class(
     args = NULL, # Save 'args' argument here
     cleanup = NULL, # cleanup argument
     cleanup_tree = NULL, # cleanup_tree argument
+    cleanup_grace = NULL, # cleanup_grace argument
     stdin = NULL, # stdin argument or stream
     stdout = NULL, # stdout argument or stream
     stderr = NULL, # stderr argument or stream
@@ -862,7 +869,7 @@ process_interrupt <- function(self, private) {
 
 process_kill <- function(self, private, grace, close_connections) {
   "!DEBUG process_kill '`private$get_short_name()`', pid `self$get_pid()`"
-  ret <- chain_call(
+  ret <- chain_clean_call(
     c_processx_kill,
     private$status,
     as.numeric(grace),