+++ tests/testthat/helper.R
@@ -249,3 +249,4 @@ retry_until <- function(fn, interrupt = 0.2, timeout = 5) {
   skip_on_cran()
   stop("timeout")
 }
+